Split `RuntimeType` and `CargoDependency` for server code generation (#870)
For clear separation of concerns.
