+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/rustlang/CargoDependency.kt
@@ -189,10 +189,8 @@ data class CargoDependency(
     }
 
     companion object {
-        val Axum: CargoDependency = CargoDependency("axum", CratesIo("0.3"))
         val Bytes: CargoDependency = CargoDependency("bytes", CratesIo("1"))
         val BytesUtils: CargoDependency = CargoDependency("bytes-utils", CratesIo("0.1.1"))
-        val DeriveBuilder = CargoDependency("derive_builder", CratesIo("0.10"))
         val FastRand: CargoDependency = CargoDependency("fastrand", CratesIo("1"))
         val Hex: CargoDependency = CargoDependency("hex", CratesIo("0.4.3"))
         val HttpBody: CargoDependency = CargoDependency("http-body", CratesIo("0.4"))
